Move deprecation_warning / deprecation_error into utils
