+++ modal/cli/app.py
@@ -10,9 +10,9 @@ from rich.text import Text
 from typer import Argument
 
 from modal._utils.async_utils import synchronizer
+from modal._utils.deprecation_utils import deprecation_warning
 from modal.client import _Client
 from modal.environments import ensure_env
-from modal.exception import deprecation_warning
 from modal.object import _get_environment_name
 from modal_proto import api_pb2
