+++ modal/config.py
@@ -86,8 +86,9 @@ from google.protobuf.empty_pb2 import Empty
 
 from modal_proto import api_pb2
 
+from ._utils.deprecation_utils import deprecation_error
 from ._utils.logger import configure_logger
-from .exception import InvalidError, deprecation_error
+from .exception import InvalidError
 
 # Locate config file and read it
