+++ modal/functions.py
@@ -40,7 +40,7 @@ from ._utils.async_utils import (
     synchronizer,
     warn_if_generator_is_not_consumed,
 )
-from ._utils.deprecation_utils import renamed_parameter
+from ._utils.deprecation_utils import deprecation_warning, renamed_parameter
 from ._utils.function_utils import (
     ATTEMPT_TIMEOUT_GRACE_PERIOD,
     OUTPUTS_TIMEOUT,
@@ -64,7 +64,6 @@ from .exception import (
     InvalidError,
     NotFoundError,
     OutputExpiredError,
-    deprecation_warning,
 )
 from .gpu import GPU_T, parse_gpu_config
 from .image import _Image
