+++ modal/partial_function.py
@@ -15,9 +15,10 @@ import typing_extensions
 from modal_proto import api_pb2
 
 from ._utils.async_utils import synchronize_api, synchronizer
+from ._utils.deprecation_utils import deprecation_error, deprecation_warning
 from ._utils.function_utils import callable_has_non_self_non_default_params, callable_has_non_self_params
 from .config import logger
-from .exception import InvalidError, deprecation_error, deprecation_warning
+from .exception import InvalidError
 from .functions import _Function
 
 MAX_MAX_BATCH_SIZE = 1000
