+++ modal/serving.py
@@ -11,12 +11,12 @@ from synchronicity.async_wrap import asynccontextmanager
 from modal._output import OutputManager
 
 from ._utils.async_utils import TaskContext, asyncify, synchronize_api, synchronizer
+from ._utils.deprecation_utils import deprecation_error
 from ._utils.logger import logger
 from ._watcher import watch
 from .cli.import_refs import import_app
 from .client import _Client
 from .config import config
-from .exception import deprecation_error
 from .output import _get_output_manager, enable_output
 from .runner import _run_app, serve_update
